updates to new db instructions
