+++ database_for_new_deployment/restore_iprdb_dump.sh
@@ -62,6 +62,9 @@ psql -U $POSTGRES_USER  -c "CREATE ROLE $READONLY_USER; ALTER ROLE $READONLY_USE
 # create the template database
 psql -U $POSTGRES_USER -c "CREATE DATABASE $TEMPLATE_NAME OWNER $SERVICE_USER IS_TEMPLATE = true;"
 psql -U $POSTGRES_USER -c "GRANT CONNECT ON DATABASE $TEMPLATE_NAME TO PUBLIC; REVOKE TEMPORARY ON DATABASE $TEMPLATE_NAME FROM PUBLIC;"
+psql -U $POSTGRES_USER -d "$TEMPLATE_NAME" -c "CREATE EXTENSION IF NOT EXISTS \"uuid-ossp\";"
+psql -U $POSTGRES_USER -d "$TEMPLATE_NAME" -c "CREATE EXTENSION IF NOT EXISTS \"fuzzystrmatch\";"
+psql -U $POSTGRES_USER -d "$TEMPLATE_NAME" -c "CREATE EXTENSION IF NOT EXISTS \"pg_trgm\";"
 
 # import dump
 PGPASSWORD=$SERVICE_PASSWORD createdb -U $SERVICE_USER -T $TEMPLATE_NAME $DATABASE_NAME
